usmsb-sdk/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "usmsb-sdk"
version = "0.1.0"
description = "USMSB (Universal System Model of Social Behavior) SDK for AI Agent Platform"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Felix Gu", email = "felix@usmsb.ai"}
]
keywords = ["ai", "agent", "llm", "usmsb", "social-behavior", "sdk"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9",
    "httpx>=0.25",
    "openai>=1.0",
    "google-generativeai>=0.3",
    "langchain>=0.1",
    "langchain-openai>=0.0.5",
    "fastapi>=0.109",
    "uvicorn>=0.27",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.19",
    "redis>=5.0",
    "structlog>=24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "black>=23.0",
    "isort>=5.0",
    "mypy>=1.0",
    "ruff>=0.1",
]
vector-db = [
    "pinecone-client>=2.2",
    "weaviate-client>=3.25",
]
graph-db = [
    "neo4j>=5.15",
]

[project.scripts]
usmsb = "usmsb_sdk.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP", "B", "C4"]
